[REFACTO] deletion of dataService. Creation of listProjectService
